+++ seacatauth/openidconnect/handler/introspect.py
@@ -5,7 +5,7 @@ import aiohttp.web
 import asab
 import asab.web.rest
 
-from ...generic import nginx_introspection, get_bearer_token_value
+from ...generic import nginx_introspection, get_bearer_token_value, get_access_token_value_from_websocket
 
 #
 
@@ -82,7 +82,9 @@ class TokenIntrospectionHandler(object):
 	async def _authenticate_request(self, request):
 		token_value = get_bearer_token_value(request)
 		if token_value is None:
-			L.log(asab.LOG_NOTICE, "No Bearer token in Authorization header.")
+			token_value = get_access_token_value_from_websocket(request)
+		if token_value is None:
+			L.log(asab.LOG_NOTICE, "Access token not found in 'Authorization' nor 'Sec-WebSocket-Protocol' header")
 			return None
 		session = await self.OpenIdConnectService.get_session_by_access_token(token_value)
 		if session is None: